requirements.txt
astroid==1.6.3
certifi==2018.4.16
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
defusedxml==0.5.0
Django==2.0.4
django-allauth==0.36.0
django-cors-headers==2.2.0
django-filter==1.1.0
django-rest-auth==0.9.3
django-rest-swagger==2.2.0
django-sslserver==0.20
djangorestframework==3.8.2
idna==2.6
isort==4.3.4
itypes==1.1.0
Jinja2==2.10
lazy-object-proxy==1.3.1
Markdown==2.6.11
MarkupSafe==1.0
mccabe==0.6.1
oauthlib==2.0.7
openapi-codec==1.3.2
pkg-resources==0.0.0
pylint==1.8.4
python3-openid==3.1.0
pytz==2018.4
requests==2.18.4
requests-oauthlib==0.8.0
simplejson==3.15.0
six==1.11.0
uritemplate==3.0.0
urllib3==1.22
wrapt==1.10.11
